requirements.txt
aiohttp==3.2.1
pyyaml
aiohttp_jinja2